requirements.txt
aioredis==1.3.1
asgi-redis==1.4.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
cffi==1.14.3
channels==3.0.2
channels-redis==3.2.0
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.0
Django==3.1.1
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
msgpack==1.0.0
msgpack-python==0.5.6
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2020.4
pyzbar==0.1.8
redis==2.10.6
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.4.1
zope.interface==5.2.0
